requirements.txt
absl-py==0.11.0
altgraph==0.17
astroid==2.4.2
astunparse==1.6.3
beautifulsoup4==4.9.3
bleach==1.5.0
bs4==0.0.1
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
colorama==0.4.4
decorator==4.4.2
Flask==1.1.2
flatbuffers==1.12
future==0.18.2
gast==0.3.3
geocoder==1.38.1
geographiclib==1.50
geopy==2.1.0
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
gTTS==2.2.1
h5py==2.10.0
html5lib==0.9999999
idna==2.10
importlib-metadata==3.1.0
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
Keras-Preprocessing==1.1.2
keyboard==0.13.5
lazy-object-proxy==1.4.3
Logbook==1.5.3
lxml==4.6.2
Markdown==3.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
multipledispatch==0.6.0
nltk==3.5
numpy==1.19.4
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==0.22.0
pandas-datareader==0.8.1
patsy==0.5.1
pefile==2019.4.18
Pillow==8.0.1
playsound==1.2.2
polygon-api-client==0.1.9
protobuf==3.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pylint==2.6.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-interface==1.6.0
pytz==2021.1
pywin32==300
pywin32-ctypes==0.2.0
ratelim==0.1.6
regex==2020.11.13
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.5.4
simplecalculator==0.0.4
simplejson==3.4.1
six==1.15.0
soupsieve==2.1
SpeechRecognition==3.8.1
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
tflearn==0.5.0
toml==0.10.2
toolz==0.11.1
tqdm==4.54.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
Wikipedia-API==0.5.4
wrapt==1.12.1
zipp==3.4.0
